app/requirements.txt
Flask==3.0.0
Pillow==9.4.0
scipy==1.10.1
tensorflow==2.13.0rc1
numpy==1.24.2
pandas==1.5.3
psycopg2-binary == 2.9.9
gunicorn == 21.2.0
flask-simplelogin==0.1.2
Flask-WTF==1.2.1